added routers, refactored main
